requirements.txt
langchain
langchain_community
langchain_groq
langchain_huggingface
chromadb
python-dotenv
tiktoken
streamlit